Subscribe to value events for user task changes
Rather than doing a database.ref().once("value") on each mount of the
TasksList page, do a database.ref().on("value", ...) on mount and a
corresponding off("value, ...) on unmount. This solves the problem of
needing to wait for the firebase cloud function to populate the user's
data on first login before doing the initial fetch of the tasks list.